front-end/src/test/javascript/spec/app/entities/ApiMonitoramentoBarragens/barragem/barragem-update.component.spec.ts
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { FormBuilder } from '@angular/forms';
import { of } from 'rxjs';

import { FrontEndTestModule } from '../../../../test.module';
import { BarragemUpdateComponent } from 'app/entities/ApiMonitoramentoBarragens/barragem/barragem-update.component';
import { BarragemService } from 'app/entities/ApiMonitoramentoBarragens/barragem/barragem.service';
import { Barragem } from 'app/shared/model/ApiMonitoramentoBarragens/barragem.model';

describe('Component Tests', () => {
  describe('Barragem Management Update Component', () => {
    let comp: BarragemUpdateComponent;
    let fixture: ComponentFixture<BarragemUpdateComponent>;
    let service: BarragemService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [FrontEndTestModule],
        declarations: [BarragemUpdateComponent],
        providers: [FormBuilder]
      })
        .overrideTemplate(BarragemUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(BarragemUpdateComponent);
      comp = fixture.componentInstance;
      service = fixture.debugElement.injector.get(BarragemService);
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', fakeAsync(() => {
        // GIVEN
        const entity = new Barragem(123);
        spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.update).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));

      it('Should call create service on save for new entity', fakeAsync(() => {
        // GIVEN
        const entity = new Barragem();
        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.create).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));
    });
  });
});
